requirements.txt
appnope==0.1.4
asgiref==3.5.2
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.5
django-cleanup==8.1.0
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.15.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
idna==3.3
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.12.51
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.14.1
Pygments==2.18.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pyzmq==26.0.3
qrcode==7.4.2
referencing==0.35.1
requests==2.28.1
rpds-py==0.18.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.3
sw-python-utils==0.0.18
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
